pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-auto-continue-macos"
version = "0.7.4"
description = "Automatically click the 'Continue' button in the Claude macOS desktop app when tool-use limits pause your session."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "Ryan S." },
]
keywords = ["claude", "anthropic", "macos", "accessibility", "automation", "electron"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "pyobjc-core>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "rich>=13.0",
    "tomli>=2.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
claude-auto-continue = "claude_auto_continue.cli:main"

[project.urls]
Homepage = "https://github.com/PrivateVictories-Main/claude-auto-continue-macos"
Issues = "https://github.com/PrivateVictories-Main/claude-auto-continue-macos/issues"
Changelog = "https://github.com/PrivateVictories-Main/claude-auto-continue-macos/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["claude_auto_continue*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
claude_auto_continue = ["dashboard_ui.html", "py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "N"]
ignore = [
    "E501",    # line length handled by formatter
    "N802",    # stdlib naming (do_GET, do_POST)
    "N803",    # argument naming (AX constants)
    "N806",    # dynamic class naming (BoundHandler)
    "UP007",   # Optional[X] vs X | None — keep for 3.9 compat
    "UP012",   # unnecessary encoding="utf-8" — explicit is fine
    "UP035",   # import from collections.abc — keep typing imports for clarity
    "UP045",   # X | None — keep Optional for 3.9 readability
    "SIM105",  # contextlib.suppress — try/except/pass is clearer in hot paths
    "SIM102",  # nested if — sometimes more readable than combined
    "SIM112",  # __CFBundleIdentifier is a macOS constant, not our naming
]

[tool.ruff.lint.isort]
known-first-party = ["claude_auto_continue"]
